possible type error getting analysis vat amount

+++ bika/lims/content/analysis.py
@@ -428,13 +428,13 @@ class Analysis(BaseContent):
                       Decimal(price) * Decimal(priority.getPricePremium()) \
                       / 100)
         return price
-         
+
     def getVATAmount(self):
         vat = self.getService().getVAT()
         price = self.getPrice()
-        return price * Decimal(vat) / 100
-         
+        return float(price) * float(vat) / 100
+
     def getTotalPrice(self):
-        return self.getPrice() + self.getVATAmount()
+        return float(self.getPrice()) + float(self.getVATAmount())
 
 atapi.registerType(Analysis, PROJECTNAME)
